Bad token of activity when do the launch work at the app process side
If the activity destroy timeout, AMS will clear it from the task.

And if activity LAUNCH_ACTIVITY message is blocked in the app main thread,
it may do the handleLaunchActivity after AMS removed the activity record
token from the task.At this time, ActivityThread launch the activity and
call some AMS interfaces like reportSizeConfigurations(isTopOfTask etc.),
AMS may throw IllegalArgumentException for the bad token.

Good for the system stability, we should check if the launching activity
is in any stack to deceide whether need to do the follow-up launch work.

Test: monkey test 20h with no crash.

Change-Id: I508601a9e944d71c9f1955fb387b8a623f0aa8fd

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>